More tagBox work
- Several bug fixes, performance improvements within SharePoint
service, added a lot of code commenting as well getting ready to
publish Angular-SharePoint module open source

- Removed LZ-String and replaced with RawDeflate, considerable
performance gain on large strings with little increase in storage size

- Major customizations of SelectizeJS:

  - Several style updates including some CSS animation, larger headers,
gradient backgrounds on columns for readability

  - Promoted input field to independent control to make search buildups
more logical

  - Fixed annoying scrollTop bug on filtering/unfiltering large lists

  - Re-patched selectize.js and optgroup_columns.js with previous fixes
for performance and dynamic column resizing

  - Removed several minor/annoying performance penalties within
SelectizeJS not needed for FTSS

  - Changed tags to float:right with input float:left so that the input
always comes first, also patched SelectizeJS to work with this

- Major tagbox function change: no longer does live searching after X #
of milliseconds.  This is a convenient feature but for the typical use
case would actually be very cumbersome and slow.  Now, searches will be
built and then fired on ENTER or Blur event.  Pending tags will appear
in a bold green font to show they have not been loaded yet.  This makes
it much more natural to build up a query and save it for future
use/sharing.

- Fixed incorrect CACHE_COUNT on dropdown init for TagBox

- Started some experimenting with CSS3 animations for a smoother UX