Milestone: finally implemented complete drag & drop/ auto-upload solution for instructor bio photos
Bio photos are now seamlessly updated by dragging & dropping the file
onto the view.  The content is sent SP and the new image loaded using
the _w pre-generated image from SP.  Caching is not as good as a
long-expiry (still uses etag) but at least it has that.

jQuery $.ajax function/ $timeout was used instead of $http due to
issues with SP accepting the ByteArray from it.