Added utils.alert{} w/three functions: create(), update(), error(), standardized alerting system using angular-strap
An edge case exists where a user could create a record or edit a record
and cause it to not show after add/update on the main view.  For
example, if a text filter is set for some particular string and the new
record did not have that string, the record would create successfully
but the user would have no real way of knowing.

Now, all save actions show a spinner to indicate work-in-progress and
then either a success or failure alert that automatically closes after
3 seconds.

This will likely be controlled through the user preference pane later
on.