Stripped Id from model--coded into AngulareSharePoint
Since Id is essential for caching (and it was causing major headaches
when I forgot to specify it) it is now automatically added to the
$select parameter if $select is specified and Id not listed.  This will
eleminate a really nasty cache corruption bug.  Basically, when no Id
was requested, an arbitrary index was used for each record causing
overwrites of previous caches—that’s very bad.  This only occurred when
new models were added without selecting an Id field but the bug was bad
enough that the SP library should prevent this if at all possible.