Added automatic removal of submitted students from backlog
This was an unfinished issue with the AAA import version of FTSS.
Basically, the AAA is cached for the session and so submitted students
would still show up on the backlog until you refreshed the page and
re-imported the AAA.

Now, we cache the unparsed AAA text and reparse everything after each
898 submission.  This allows us to check the updated RequirementStats
model for changes and reflect them in the view.

For the user, this means when you submit an 898, even if the AAA still
reflects the students as AWACT/AWFTD, they will not show up on the
backlog but instead will be on the previously requested list under
Inspect Evaluated AAA Data.